requirements.txt
Langchain==0.0.154
pyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.7.4
streamlit-extras
openai
